requirements.txt.txt
django
djangorestframework
django-filter
markdown
Pillow
dj_rest_auth
corsheaders
djangorestframework-simplejwt